+++ hyperdiary/check.py
@@ -1,15 +1,11 @@
 from datetime import date, timedelta
 
 def check(diary):
-    current = date(2016,1,5)
-    today = date.today()
-    one_day = timedelta(days=1)
+    dates_found = set()
+    for dtrange in diary.expected:
+        for dt in dtrange:
+            if not dt in diary.entries:
+                raise Exception('Missing entry {}'.format(dt))
+            dates_found.add(dt)
 
-    number_of_days = 0
-    while current < today:
-        if not current in diary.entries:
-            raise Exception('Missing entry {}'.format(current))
-        current += one_day
-        number_of_days += 1
-
-    print('OK found {} days'.format(number_of_days))
+    print('OK found {} days'.format(len(dates_found)))