+++ tests/test_hyperdiary.py
@@ -8,6 +8,7 @@ from tempfile import TemporaryDirectory
 
 from hyperdiary import parser, Diary
 from hyperdiary.diary import find_ids, find_tags, tokenize
+from hyperdiary.check import check
 from hyperdiary.html import diary_to_html, diary_to_html_folder
 from hyperdiary.tiddlywiki import diary_to_tiddlers
 from hyperdiary.hugo import diary_to_hugo
@@ -28,6 +29,9 @@ class TestHyperdiary(unittest.TestCase):
     
     def test_loading_of_entries(self):
         self.assertGreaterEqual(len(self.diary.entries), 3)
+    
+    def test_check(self):
+        check(self.diary)
     
     def test_missing_hyperdiary_json(self):
         self.assertRaises(FileNotFoundError, Diary.discover,