+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.3.0 (not yet)
 ---------------
 
+* Expected date ranges for check subcommand can be specified
 * Tests for various output formats
 
 0.2.0 (2019-06-12)